etl/requirements.txt
pandas
sqlalchemy
psycopg2-binary
spotipy
python-dotenv
musicbrainzngs